seperate cache functions to cache module, export/import setCache(), load main.js as type module
