+++ index.html
@@ -56,7 +56,7 @@
       </div>
     </div>
   </div>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js" ></script>
 </body>
 
 </html>
